Move surject into a proper subcommand
